refactor: promote createActiveSource to @real-router/sources (#1416)
The #1416 fix mirrored the angular createActiveSource helper into
packages/vue/src/internal/ — a faithful copy, which SonarCloud flagged as 81.8%
duplicated new code (the fast/slow builder is framework-agnostic, so the vue and
angular copies were near-identical).

Promote the builder to @real-router/sources, where it belongs (it uses only
createActiveNameSelector + createActiveRouteSource). vue <Link>/useIsActiveRoute
and angular RealLink/RealLinkActive now import createActiveSource from sources;
both per-adapter copies (and their duplicated tests) are deleted, the helper test
is consolidated in sources. One source of truth for the fast/slow decision (+ the
routeName !== "" guard) across all adapters — the strongest form of the anti-drift
goal #1416 was about.

No behavior change: the builder's logic + signature are unchanged. sources + vue +
angular changesets; sources 268 / vue 466 / angular 395 tests pass, type-check +
lint clean.
