+++ packages/angular/src/directives/RealLink.ts
@@ -7,10 +7,10 @@ import {
   input,
   signal,
 } from "@angular/core";
+import { createActiveSource } from "@real-router/sources";
 
 import { buildHref, navigateWithHash, shouldNavigate } from "../dom-utils";
 import { injectRouter } from "../functions/injectRouter";
-import { createActiveSource } from "../internal/createActiveSource";
 import { createStableParams } from "../internal/createStableParams";
 import { subscribeSourceToSignal } from "../internal/subscribeSourceToSignal";
