+++ packages/sources/tests/functional/createActiveSource.test.ts
@@ -1,24 +1,20 @@
 import { createRouter } from "@real-router/core";
-import { createActiveRouteSource } from "@real-router/sources";
 import { describe, beforeEach, afterEach, it, expect, vi } from "vitest";
 
-import { createActiveSource } from "../../src/internal/createActiveSource";
+import { createActiveRouteSource, createActiveSource } from "../../src";
 
 import type { Router } from "@real-router/core";
 
-// `createActiveSource` is the fast/slow active-source builder shared by the
-// `<Link>` reactive `watch` and the `useIsActiveRoute` composable (#1416). It
-// takes the router explicitly, so it is directly unit-testable — pinning the
-// path decision + the fast-path reactive contract in ONE place (the fast/slow
-// drift between `<Link>` and `useIsActiveRoute` is exactly what produced #1416:
-// #1250 landed the fast path in the composable, `<Link>` kept its own slow path).
+// `createActiveSource` is the framework-agnostic fast/slow active-source builder
+// shared by every adapter's `<Link>` (#1416 promoted it here from the per-adapter
+// copies so the fast/slow decision cannot drift).
 //
 // Discriminator for "fast path was taken": a fast-path source is backed by the
 // shared `createActiveNameSelector` and never builds the per-link
 // `createActiveRouteSource`, so the canonical slow-path source for the same
 // arguments is still UNBUILT — asking for it is a cache MISS that runs
-// `router.isActiveRoute` once. A slow-path `createActiveSource` builds that
-// exact source, so the same later call is a cache HIT (no `isActiveRoute`).
+// `router.isActiveRoute` once. A slow-path `createActiveSource` builds that exact
+// source, so the same later call is a cache HIT (no `isActiveRoute`).
 describe("createActiveSource", () => {
   const routes = [
     { name: "home", path: "/" },
@@ -154,8 +150,7 @@ describe("createActiveSource", () => {
     // routeName="" is a misuse (no href, console.error). It must NOT take the
     // fast path — the selector reports root-active `true` for "", whereas the
     // slow path's isActiveRoute("") is false. Keeping "" on the slow path
-    // preserves that behaviour (the pre-fix `useIsActiveRoute` fast path lacked
-    // this guard — a latent bug the shared builder fixes for both callers).
+    // preserves that behaviour.
     createActiveSource(router, "", undefined, false, true, undefined);
 
     const spy = vi.spyOn(router, "isActiveRoute");
@@ -170,15 +165,11 @@ describe("createActiveSource", () => {
     spy.mockRestore();
   });
 
-  it("paramless link to a param route is name-only active while a param instance is active (#1416 — aligns vue with the other adapters)", async () => {
-    // Ship-gate 2 (#1416): a paramless `createActiveSource(router, "users.view")`
-    // takes the FAST path (name-only). While `users.view` is active with a
-    // concrete `{ id }`, the selector reports `isActive("users.view") === true`
-    // by NAME — it ignores the params. Pre-fix, vue's `<Link>` built an
-    // undefined-params slow-path source (compared via `areStatesEqual`) that
-    // returned `false` for this case; the fast path aligns vue with the react /
-    // preact / solid / svelte / angular adapters, all name-only for a paramless
-    // link.
+  it("paramless link to a param route is name-only active while a param instance is active (#1416)", async () => {
+    // A paramless `createActiveSource(router, "users.view")` takes the FAST path
+    // (name-only). While `users.view` is active with a concrete `{ id }`, the
+    // selector reports `isActive("users.view") === true` by NAME — it ignores the
+    // params. Every adapter Link is name-only for a paramless link.
     await router.navigate("users.view", { id: "5" });
 
     const source = createActiveSource(
